@@ -20,7 +20,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 3.0.0+13
 
 environment:
-  sdk: ">=2.18.1 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -31,7 +31,7 @@ environment:
 dependencies:
   auto_size_text: ^3.0.0
   cached_network_image: ^3.4.1
-  carousel_slider: ^4.2.1
+  carousel_slider: ^5.0.0
   cloud_firestore: ^4.13.6
   collection: any
   cupertino_icons: ^1.0.2
@@ -64,7 +64,12 @@ dependencies:
   timelines: ^0.1.0
   url_launcher: ^6.3.0
   xml: null
+  win32: ^5.9.0
+  drift: ^2.22.1
+  drift_flutter: ^0.2.3
 dev_dependencies:
+  build_runner: ^2.4.13
+  drift_dev: ^2.22.1
   flutter_launcher_icons: ^0.13.1
   flutter_lints: ^3.0.1
   flutter_test:
