@@ -9,6 +9,7 @@ import 'package:zineapp2023/screens/chat/chat_screen/view_model/chat_room_view_m
 import 'package:zineapp2023/screens/dashboard/view_models/dashboard_vm.dart';
 import 'package:zineapp2023/theme/color.dart';
 import '../../../components/gradient.dart';
+import '../../../database/database.dart';
 import '../../../models/rooms.dart';
 import 'chat_view.dart';
 
@@ -32,11 +33,12 @@ class _ChatRoomState extends State<ChatRoom> {
     _saveRoomNameToPreferences();
     WidgetsBinding.instance.addPostFrameCallback((_) {
       chatRoomView = Provider.of<ChatRoomViewModel>(context, listen: false);
+      var db = Provider.of<AppDb>(context, listen: false);
       widget.roomDetail?.id != null
-          ? chatRoomView.fetchMessages(widget.roomDetail!.id.toString())
+          ? chatRoomView.staticMessagePipeline(db,widget.roomDetail!.id.toString())//chatRoomView.fetchMessages(widget.roomDetail!.id.toString())
           : "";
       widget.roomDetail?.id != null
-          ? chatRoomView.setRoomId(widget.roomDetail!.id.toString())
+          ? chatRoomView.setRoomId(widget.roomDetail!.id.toString(),db)
           : "";
       chatRoomView.getTotalActiveMember(widget.roomDetail!.id.toString());
     });
